+++ amd64/README.md
@@ -39,13 +39,13 @@ The package API documentation can be found on
 Supported hardware
 ==================
 
-| CPU              | Related platform packages                                                                        | Core drivers                                                         | Peripheral drivers                                                                                                                                                                  |
-|------------------|--------------------------------------------------------------------------------------------------|----------------------------------------------------------------------|-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|
-| AMD/Intel 64-bit | [cloud_hypervisor/vm](https://github.com/usbarmory/tamago/tree/master/board/cloud_hypervisor/vm) | [LAPIC](https://github.com/usbarmory/tamago/tree/master/amd64/lapic) | [KVM clock, VirtIO over PCI](https://github.com/usbarmory/tamago/tree/master/kvm), [IOAPIC, PCI, RTC, UART](https://github.com/usbarmory/tamago/blob/master/soc/intel)              |
-| AMD/Intel 64-bit | [qemu/microvm](https://github.com/usbarmory/tamago/tree/master/board/qemu/microvm)               | [LAPIC](https://github.com/usbarmory/tamago/tree/master/amd64/lapic) | [KVM clock, VirtIO over MMIO](https://github.com/usbarmory/tamago/tree/master/kvm), [IOAPIC, RTC, UART](https://github.com/usbarmory/tamago/blob/master/soc/intel)                  |
-| AMD/Intel 64-bit | [firecracker/microvm](https://github.com/usbarmory/tamago/tree/master/board/firecracker/microvm) | [LAPIC](https://github.com/usbarmory/tamago/tree/master/amd64/lapic) | [KVM clock, VirtIO over MMIO](https://github.com/usbarmory/tamago/tree/master/kvm), [IOAPIC, UART](https://github.com/usbarmory/tamago/blob/master/soc/intel)                       |
-| AMD/Intel 64-bit | [google/gcp](https://github.com/usbarmory/tamago/tree/master/board/google/gcp)                   | [LAPIC](https://github.com/usbarmory/tamago/tree/master/amd64/lapic) | [KVM clock, VirtIO over PCI](https://github.com/usbarmory/tamago/tree/master/kvm), [IOAPIC, PCI, RTC, UART](https://github.com/usbarmory/tamago/blob/master/soc/intel)              |
-| AMD/Intel 64-bit | [uefi/x64](https://github.com/usbarmory/go-boot/tree/main/uefi/x64)                              |                                                                      | [EFI Console I/O, Graphics, Boot and Runtime Services](https://github.com/usbarmory/go-boot/tree/main/uefi), [RTC, UART](https://github.com/usbarmory/tamago/blob/master/soc/intel) |
+| CPU              | Related platform packages                                                                        | Core drivers                                                                                                                         | Peripheral drivers                                                                                                                                                     |
+|------------------|--------------------------------------------------------------------------------------------------|--------------------------------------------------------------------------------------------------------------------------------------|------------------------------------------------------------------------------------------------------------------------------------------------------------------------|
+| AMD/Intel 64-bit | [cloud_hypervisor/vm](https://github.com/usbarmory/tamago/tree/master/board/cloud_hypervisor/vm) | [LAPIC](https://github.com/usbarmory/tamago/tree/master/amd64/lapic)                                                                 | [KVM clock, VirtIO over PCI](https://github.com/usbarmory/tamago/tree/master/kvm), [IOAPIC, PCI, RTC, UART](https://github.com/usbarmory/tamago/blob/master/soc/intel) |
+| AMD/Intel 64-bit | [qemu/microvm](https://github.com/usbarmory/tamago/tree/master/board/qemu/microvm)               | [LAPIC](https://github.com/usbarmory/tamago/tree/master/amd64/lapic), [SEV](https://github.com/usbarmory/tamago/tree/master/kvm/sev) | [KVM clock, VirtIO over MMIO](https://github.com/usbarmory/tamago/tree/master/kvm), [IOAPIC, RTC, UART](https://github.com/usbarmory/tamago/blob/master/soc/intel)     |
+| AMD/Intel 64-bit | [firecracker/microvm](https://github.com/usbarmory/tamago/tree/master/board/firecracker/microvm) | [LAPIC](https://github.com/usbarmory/tamago/tree/master/amd64/lapic)                                                                 | [KVM clock, VirtIO over MMIO](https://github.com/usbarmory/tamago/tree/master/kvm), [IOAPIC, UART](https://github.com/usbarmory/tamago/blob/master/soc/intel)          |
+| AMD/Intel 64-bit | [google/gcp](https://github.com/usbarmory/tamago/tree/master/board/google/gcp)                   | [LAPIC](https://github.com/usbarmory/tamago/tree/master/amd64/lapic), [SEV](https://github.com/usbarmory/tamago/tree/master/kvm/sev) | [KVM clock, VirtIO over PCI](https://github.com/usbarmory/tamago/tree/master/kvm), [IOAPIC, PCI, RTC, UART](https://github.com/usbarmory/tamago/blob/master/soc/intel) |
+| AMD/Intel 64-bit | [uefi/x64](https://github.com/usbarmory/go-boot/tree/main/uefi/x64)                              | [SEV](https://github.com/usbarmory/tamago/tree/master/kvm/sev)                                                                       | [EFI Console I/O, Graphics, Boot and Runtime Services](https://github.com/usbarmory/go-boot/tree/main/uefi)                                                            |
 
 Build tags
 ==========
