+++ kvm/sev/msg_test.go
@@ -1,4 +1,4 @@
-// AMD secure virtualization support
+// AMD Secure Encrypted Virtualization support
 // https://github.com/usbarmory/tamago
 //
 // Copyright (c) The TamaGo Authors. All Rights Reserved.
@@ -6,7 +6,7 @@
 // Use of this source code is governed by the license
 // that can be found in the LICENSE file.
 
-package svm
+package sev
 
 import (
 	"bytes"
